Extract posterior samples for a single parameter

Wrapper for extract() applied to the stanfit component of a fitted salmonIPMfit object. Unlike extract(), only returns the first element, not a named list. This is useful for obtaining the samples for a single parameter (of any dimension), though generally as.matrix(object, par) is preferred.

Usage

extract1(object, par)

Arguments

object

An object of class salmonIPMfit.

par

Character string giving quantity to return.