Pointwise log-likelihood matrix

The log_lik method for salmonIPMfit objects returns the pointwise log-likelihood matrix.

Usage

# S3 method for class 'salmonIPMfit'
log_lik(object, ...)

# S3 method for class 'salmonIPMfit'
nsamples(object, ...)

Arguments

object

An object of class salmonIPMfit that includes samples of the pointwise log-likelihood LL, e.g. log_lik was set to TRUE when calling salmonIPM().

...

Currently ignored.

Value

A N_samples by N matrix containing the pointwise log-likelihood, where N_samples is the number of post-warmup posterior draws and N is the number of cases (i.e. rows) in fish_data. Note that this includes all likelihood components for observations in fish_data, but does not include the likelihood of auxiliary data types. These, as well as the likelihood components for specific data types in fish_data, can be extracted with as.matrix() if they have been monitored using the pars argument to salmonIPM().

See also