Prior distributions

These functions are used to specify priors on selected (hyper)parameters in salmonIPM models.

The default priors used in the various models are intended to be weakly informative, in that they provide moderate regularization and help stabilize sampling. Priors on scaling parameters, e.g. Rmax or mu_Rmax, are automatically adjusted to be weakly informative but consistent with the observed marginal distribution of population density. For many applications these defaults will perform well, but if external information not included in fish_data is available, it can be incorporated via user-specified priors on key parameters. See Details for a table of available prior options.

Usage

normal(mean = 0, sd = 1)

gnormal(mean = 0, scale = 1, shape = 1)

lognormal(meanlog = 0, sdlog = 1)

beta(a = 1, b = 1)

dirichlet(concentration = 1)

uniform(lb = 0, ub = 1)

lkj_corr(eta = 1)

Arguments

mean

Prior mean for normal or generalized normal distribution.

sd

Prior standard deviation for normal distribution.

scale

Prior scale for generalized normal distribution. Equivalent to alpha in gnorm, but renamed to avoid confusion with the spawner-recruit intrinsic productivity parameter.

shape

Prior shape for generalized normal distribution. Equivalent to beta in gnorm but renamed to avoid confusion with covariate slopes.

meanlog, sdlog

Prior log-scale mean and standard deviation, respectively, for lognormal distribution. See Lognormal.

a, b

Prior shape parameters for the beta distribution. Equivalent to shape1 and shape2, respectively, in Beta.

concentration

Vector of shape parameters for the Dirichlet distribution. Equivalent to alpha in gtools::dirichlet, but renamed to avoid confusion with the spawner-recruit intrinsic productivity parameter.

lb, ub

Lower and upper bounds for the uniform distribution.

eta

Prior shape parameter for the LKJ distribution over correlation matrices.

Value

A named list to be used internally by the salmonIPM model-fitting and summary functions.

Details

The table below shows the parameters in each model that can be given user-specified priors and the corresponding distributions. Note that users can modify the prior parameters but not the distribution families; attempting to do the latter will result in an error.

Priors for parameters that are bounded on the positive real line (e.g. tau, tau_S and tau_M) are automatically left-truncated at zero.

For parameters that are modeled as functions of covariates using the par_models argument to salmonIPM(), the specified prior applies when all predictors are at their sample means.

If RRS != "none", the global spawner-recruit parameters must be replaced with their W and H counterparts; e.g. if RRS == "alpha" then instead of a prior on alpha one would specify priors on alpha_W and alpha_H. If the former is provided, it will have no effect. See salmonIPM() for details of the RRS argument.

The generalized normal density with shape >> 1 is useful as a platykurtic "soft-uniform" prior to regularize the posterior away from regions of parameter space that may cause computational or sampling problems. In the case of spawner and smolt observation error log-SDs, the default prior bounds them ≳ 0.1.

The uniform distribution and the LKJ distribution are included for internal use; currently no correlation matrices have user-specified priors.

Parameter (PDF)

Model	alpha	lognormal	mu_alpha	normal	mu_psi	beta	Rmax	lognormal	mu_Rmax	normal	Mmax	lognormal

mu_Mmax	normal	mu_MS	beta	mu_p	dirichlet	mu_SS	beta	tau	gnormal	tau_S	tau_M	gnormal

IPM_SS_np	☑	☐	☐	☑	☐	☐	☐	☐	☑	☐	☑	☐

IPM_SSiter_np	☑	☐	☐	☑	☐	☐	☐	☐	☑	☑	☑	☐

IPM_SS_pp	☐	☑	☐	☐	☑	☐	☐	☐	☑	☐	☑	☐

IPM_SSiter_pp	☐	☑	☐	☐	☑	☐	☐	☐	☑	☑	☑	☐

IPM_SMS_np	☑	☐	☐	☐	☐	☑	☐	☑	☑	☐	☐	☑

IPM_SMS_pp	☐	☑	☐	☐	☐	☐	☑	☑	☑	☐	☐	☑

IPM_SMaS_np	☑	☐	☐	☐	☐	☑	☐	☐	☐	☐	☐	☑

IPM_LCRchum_pp	☐	☐	☑	☐	☐	☐	☑	☑	☑	☐	☐	☐