Perform run reconstruction on brood table data

Perform run reconstruction on brood table data

Usage

run_recon(fish_data, age_F = NULL, age_B = NULL)

Arguments

fish_data

Data frame that includes the following columns, in no particular order except where noted:

pop Numeric, factor or character population ID.

year Numeric variable giving the year the fish spawned (i.e., the brood year).

A Spawning habitat size (either stream length or area). Will often be time-invariant within a population, but need not be.

S_obs Total number (not density) of all wild and hatchery-origin spawners.

n_age[min_age]_obs...n_age[max_age]_obs Multiple columns of observed spawner age frequencies (i.e., counts), where [min_age] and [max_age] are the numeral age in years (total, not ocean age) of the youngest and oldest spawners, respectively. Complex age structures such as Gilbert-Rich or maiden-kelt are not supported.

n_W_obs Observed frequency of natural-origin spawners.

n_H_obs Observed frequency of hatchery-origin spawners.

F_rate Total harvest rate (proportion) of natural-origin fish.

B_take_obs Number of adults taken for hatchery broodstock.

age_F

Logical or 0/1 vector of length N_age indicating whether each adult age is fully (non)selected by the fishery. The default is all selected. If life_cycle == "SSiter", N_age refers to the total number of maiden and repeat age classes (counting the repeat plus group as 1).

age_B

Logical or 0/1 vector of length N_age indicating whether each adult age is fully (non)selected in broodstock collection. The default is all selected. If life_cycle == "SSiter", N_age refers to the total number of maiden and repeat age classes (counting the repeat plus group as 1).

Value

A data frame with the following columns, some of which are simply replicated from fish_data:

pop See above.

year See above.

A See above.

S_obs See above.

q_age[minAge]_obs...q_age[maxAge]_obs Multiple columns of spawner age proportions corresponding to the frequencies in fish_data.

p_age_minAge...p_age_maxAge Multiple columns of recruit age proportions by brood year.

p_HOS_obs Proportion of hatchery-origin spawners.

F_rate See above.

B_take_obs See above.

R_obs Total natural-origin recruits from the brood year in each row.